backend/requirements.txt
fastapi
uvicorn
numpy
openai-whisper
librosa
scipy
